+++ phoenix-shared/src/androidMain/kotlin/fr/acinq/phoenix/db/SqlPaymentHooks.kt
@@ -2,6 +2,8 @@ package fr.acinq.phoenix.db
 
 import fr.acinq.lightning.utils.UUID
 import fr.acinq.phoenix.db.payments.CloudKitInterface
+import fr.acinq.phoenix.db.sqldelight.AppDatabase
+import fr.acinq.phoenix.db.sqldelight.PaymentsDatabase
 
 actual fun didSaveWalletPayment(id: UUID, database: PaymentsDatabase) {}
 actual fun didDeleteWalletPayment(id: UUID, database: PaymentsDatabase) {}
