+++ phoenix-shared/src/androidUnitTest/kotlin/fr/acinq/phoenix/db/SqliteChannelsDatabaseTest.kt
@@ -18,6 +18,7 @@ package fr.acinq.phoenix.db
 
 import app.cash.sqldelight.db.SqlDriver
 import app.cash.sqldelight.driver.jdbc.sqlite.JdbcSqliteDriver
+import fr.acinq.phoenix.db.sqldelight.ChannelsDatabase
 
 actual fun testChannelsDriver(): SqlDriver {
     val driver: SqlDriver = JdbcSqliteDriver(JdbcSqliteDriver.IN_MEMORY)
