+++ phoenix-shared/src/commonMain/kotlin/fr.acinq.phoenix/db/DbHooks.kt
@@ -2,6 +2,8 @@ package fr.acinq.phoenix.db
 
 import fr.acinq.lightning.utils.UUID
 import fr.acinq.phoenix.db.payments.CloudKitInterface
+import fr.acinq.phoenix.db.sqldelight.AppDatabase
+import fr.acinq.phoenix.db.sqldelight.PaymentsDatabase
 
 /**
  * Implement this function to execute platform specific code when a payment is saved to the database.
