+++ phoenix-shared/src/commonMain/kotlin/fr.acinq.phoenix/db/SqliteAppDb.kt
@@ -12,6 +12,8 @@ import fr.acinq.phoenix.data.FiatCurrency
 import fr.acinq.phoenix.data.Notification
 import fr.acinq.phoenix.db.notifications.ContactQueries
 import fr.acinq.phoenix.db.notifications.NotificationsQueries
+import fr.acinq.phoenix.db.sqldelight.AppDatabase
+import fr.acinq.phoenix.db.sqldelight.Exchange_rates
 import kotlin.collections.List
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.flow.Flow
