+++ phoenix-shared/src/commonMain/kotlin/fr.acinq.phoenix/db/SqliteChannelsDb.kt
@@ -21,6 +21,7 @@ import fr.acinq.bitcoin.ByteVector32
 import fr.acinq.lightning.CltvExpiry
 import fr.acinq.lightning.channel.states.PersistedChannelState
 import fr.acinq.lightning.db.ChannelsDb
+import fr.acinq.phoenix.db.sqldelight.ChannelsDatabase
 import kotlin.collections.List
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.withContext
