+++ phoenix-shared/src/commonMain/kotlin/fr.acinq.phoenix/db/notifications/NotificationsQueries.kt
@@ -23,7 +23,7 @@ import fr.acinq.lightning.utils.UUID
 import fr.acinq.lightning.utils.currentTimestampMillis
 import fr.acinq.phoenix.data.Notification
 import fr.acinq.phoenix.data.WatchTowerOutcome
-import fr.acinq.phoenix.db.AppDatabase
+import fr.acinq.phoenix.db.sqldelight.AppDatabase
 import fr.acinq.phoenix.db.notifications.NotificationData.Companion.encodeAsDb
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.IO
