+++ phoenix-shared/src/commonMain/kotlin/fr.acinq.phoenix/db/payments/PaymentsMetadataQueries.kt
@@ -5,7 +5,7 @@ import fr.acinq.lightning.utils.UUID
 import fr.acinq.lightning.utils.currentTimestampMillis
 import fr.acinq.phoenix.data.*
 import fr.acinq.phoenix.data.lnurl.LnurlPay
-import fr.acinq.phoenix.db.PaymentsDatabase
+import fr.acinq.phoenix.db.sqldelight.PaymentsDatabase
 import fr.acinq.phoenix.db.didUpdateWalletPaymentMetadata
 import io.ktor.http.*
