+++ phoenix-shared/src/commonMain/kotlin/fr.acinq.phoenix/db/payments/SqliteIncomingPaymentsDb.kt
@@ -23,9 +23,8 @@ import fr.acinq.lightning.db.IncomingPayment
 import fr.acinq.lightning.db.IncomingPaymentsDb
 import fr.acinq.lightning.db.LightningIncomingPayment
 import fr.acinq.lightning.db.OnChainIncomingPayment
-import fr.acinq.lightning.utils.UUID
 import fr.acinq.lightning.wire.LiquidityAds
-import fr.acinq.phoenix.db.PaymentsDatabase
+import fr.acinq.phoenix.db.sqldelight.PaymentsDatabase
 import fr.acinq.phoenix.db.didDeleteWalletPayment
 import fr.acinq.phoenix.db.didSaveWalletPayment
 import fr.acinq.phoenix.utils.extensions.deriveUUID
