+++ phoenix-shared/src/iosMain/kotlin/fr/acinq/phoenix/db/SqlPaymentHooks.kt
@@ -3,6 +3,8 @@ package fr.acinq.phoenix.db
 import fr.acinq.lightning.utils.UUID
 import fr.acinq.lightning.utils.currentTimestampMillis
 import fr.acinq.phoenix.db.payments.CloudKitInterface
+import fr.acinq.phoenix.db.sqldelight.AppDatabase
+import fr.acinq.phoenix.db.sqldelight.PaymentsDatabase
 
 
 actual fun didSaveWalletPayment(id: UUID, database: PaymentsDatabase) {
